requirements.txt
alembic==1.7.1
altair==4.1.0
amqp==2.6.1
anyio==3.3.1
apache-airflow==2.1.4
apache-airflow-providers-celery==2.0.0
apache-airflow-providers-cncf-kubernetes==2.0.3
apache-airflow-providers-docker==2.1.1
apache-airflow-providers-ftp==2.0.1
apache-airflow-providers-http==2.0.1
apache-airflow-providers-imap==2.0.1
apache-airflow-providers-sqlite==2.0.1
apispec==3.3.2
appdirs==1.4.4
appnope==0.1.2
argcomplete==1.12.3
argon2-cffi==20.1.0
asn1crypto==1.4.0
async-generator==1.10
attrs==20.3.0
Babel==2.9.1
backcall==0.2.0
backports.entry-points-selectable==1.1.1
backports.zoneinfo==0.2.1
beautifulsoup4==4.9.3
billiard==3.6.4.0
bleach==3.3.1
blinker==1.4
bs4==0.0.1
cached-property==1.5.2
cachetools==4.2.4
cattrs==1.5.0
celery==4.4.7
certifi==2020.12.5
cffi==1.14.6
chardet==4.0.0
charset-normalizer==2.0.5
click==7.1.2
clickclick==20.10.2
coala-utils==0.5.1
colorama==0.4.4
colorlog==4.8.0
commonmark==0.9.1
configparser==5.0.2
cramjam==2.5.0
croniter==1.0.15
cryptography==3.4.8
debugpy==1.4.1
decorator==5.0.9
defusedxml==0.7.1
diff-cover==6.3.5
diff-match-patch==20200713
dill==0.3.1.1
distlib==0.3.3
dnspython==1.16.0
docker==5.0.2
docutils==0.16
email-validator==1.1.3
entrypoints==0.3
fastparquet==0.7.2
filelock==3.3.2
findspark==1.4.2
Flask==1.1.4
Flask-AppBuilder==3.3.3
Flask-Babel==1.0.0
Flask-Caching==1.10.1
Flask-Cors==3.0.10
Flask-JWT-Extended==3.25.1
Flask-Login==0.4.1
Flask-OpenID==1.3.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.14.3
flower==0.9.7
fsspec==2021.11.1
google-auth==2.3.2
graphviz==0.17
great-expectations==0.13.44
gunicorn==20.1.0
h11==0.12.0
holistics==0.0a11
html5lib==1.1
httpcore==0.13.7
httpx==0.19.0
humanize==3.11.0
idna==3.2
importlib-metadata==4.8.1
importlib-resources==1.5.0
inflect==5.3.0
inflection==0.5.1
iniconfig==1.1.1
ipykernel==6.0.3
ipython==7.26.0
ipython-genutils==0.2.0
ipywidgets==7.6.3
iso8601==0.1.16
isodate==0.6.0
itsdangerous==1.1.0
jedi==0.18.0
Jinja2==2.11.3
jinja2-pluralize==0.3.0
jsonpatch==1.32
jsonpointer==2.2
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.12
jupyter-console==6.4.0
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
kombu==4.6.11
kubernetes==11.0.0
lazy-object-proxy==1.4.3
lockfile==0.12.2
lxml==4.6.3
Mako==1.1.5
Markdown==3.3.4
MarkupSafe==1.1.1
marshmallow==3.13.0
marshmallow-enum==1.5.1
marshmallow-oneofschema==3.0.1
marshmallow-sqlalchemy==0.23.1
matplotlib-inline==0.1.2
mistune==0.8.4
mock==4.0.3
nb2xls==0.1.6
nbclient==0.5.3
nbconvert==6.1.0
nbformat==5.1.3
nest-asyncio==1.5.1
networkx==2.6.2
notebook==6.4.0
numpy==1.19.5
oauthlib==3.1.1
openapi-schema-validator==0.1.5
openapi-spec-validator==0.3.1
oscrypto==1.2.1
oyaml==1.0
packaging==21.0
pandas==1.1.5
pandocfilters==1.4.3
parso==0.8.2
pathspec==0.9.0
pbr==5.7.0
pendulum==2.1.2
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==2.4.0
pluggy==0.13.1
prison==0.2.1
progressbar2==3.53.1
prometheus-client==0.8.0
prompt-toolkit==3.0.19
psutil==5.8.0
ptyprocess==0.7.0
py==1.10.0
py4j==0.10.9
pyarrow==5.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pycryptodomex==3.11.0
Pygments==2.10.0
PyJWT==1.7.1
pyOpenSSL==20.0.1
pyparsing==2.4.7
pypng==0.0.21
pyrsistent==0.18.0
pyspark==3.0.0
pytest==6.2.4
pytest-mock==3.6.1
python-daemon==2.3.0
python-dateutil==2.8.2
python-nvd3==0.15.0
python-slugify==4.0.1
python-utils==2.5.6
python3-openid==3.2.0
pytz==2021.1
pytz-deprecation-shim==0.1.0.post0
pytzdata==2020.1
PyYAML==5.4.1
pyzmq==22.1.0
qtconsole==5.1.1
QtPy==1.9.0
requests==2.26.0
requests-mock==1.9.3
requests-oauthlib==1.3.0
responses==0.13.4
rfc3986==1.5.0
rich==10.9.0
rsa==4.7.2
ruamel.yaml==0.17.17
ruamel.yaml.clib==0.2.6
scipy==1.7.3
Send2Trash==1.7.1
setproctitle==1.2.2
six==1.16.0
sniffio==1.2.0
snowflake-connector-python==2.7.1
soupsieve==2.2.1
SQLAlchemy==1.3.24
SQLAlchemy-JSONField==1.0.0
SQLAlchemy-Utils==0.37.8
sqlfluff==0.6.4
sqllineage==1.2.4
sqlparse==0.4.1
stevedore==3.5.0
swagger-ui-bundle==0.0.9
tabulate==0.8.9
tblib==1.7.0
tenacity==6.2.0
termcolor==1.1.0
terminado==0.10.1
testpath==0.5.0
text-unidecode==1.3
thrift==0.15.0
toml==0.10.2
toolz==0.11.2
tornado==6.1
tqdm==4.62.3
traitlets==5.0.5
twitterscraper==1.6.1
typing-extensions==3.10.0.2
tzdata==2021.5
tzlocal==4.1
unicodecsv==0.14.1
urllib3==1.26.6
vine==1.3.0
virtualenv==20.10.0
virtualenv-clone==0.5.7
virtualenvwrapper==4.8.4
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.2.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
WTForms==2.3.3
XlsxWriter==3.0.1
zipp==3.5.0
